+++ src/utils-dispatch.c
@@ -13,15 +13,6 @@ SEXP vctrs_class_type(SEXP x) {
   return Rf_mkString(class_type_as_str(class_type(x)));
 }
 
-// [[ include("utils.h") ]]
-bool is_record(SEXP x) {
-  enum vctrs_class_type type = class_type(x);
-  return
-    type == vctrs_class_rcrd ||
-    type == vctrs_class_posixlt ||
-    type == vctrs_class_bare_posixlt;
-}
-
 
 // [[ include("utils.h") ]]
 enum vctrs_class_type class_type(SEXP x) {
@@ -93,17 +84,10 @@ static enum vctrs_class_type class_type_impl(SEXP class) {
   }}
 
   // Now check for inherited classes
-  p = p + n - 2;
-  SEXP butlast = *p++;
-  SEXP last = *p++;
+  p = p + n - 1;
+  SEXP last = *p;
 
-  if (butlast == strings_posixlt) {
-    if (last == strings_posixt) return vctrs_class_posixlt;
-  } else if (butlast == strings_vctrs_rcrd) {
-    if (last == strings_vctrs_vctr) return vctrs_class_rcrd;
-  } else if (butlast == strings_vctrs_list_of) {
-    if (last == strings_vctrs_vctr) return vctrs_class_list_of;
-  } else if (last == strings_data_frame) {
+  if (last == strings_data_frame) {
     return vctrs_class_data_frame;
   } else if (last == strings_list) {
     return vctrs_class_list;
@@ -115,17 +99,14 @@ static enum vctrs_class_type class_type_impl(SEXP class) {
 static const char* class_type_as_str(enum vctrs_class_type type) {
   switch (type) {
   case vctrs_class_list: return "list";
-  case vctrs_class_list_of: return "list_of";
   case vctrs_class_data_frame: return "data_frame";
   case vctrs_class_bare_data_frame: return "bare_data_frame";
   case vctrs_class_bare_tibble: return "bare_tibble";
   case vctrs_class_bare_factor: return "bare_factor";
   case vctrs_class_bare_ordered: return "bare_ordered";
-  case vctrs_class_rcrd: return "rcrd";
   case vctrs_class_bare_date: return "bare_date";
   case vctrs_class_bare_posixct: return "bare_posixct";
   case vctrs_class_bare_posixlt: return "bare_posixlt";
-  case vctrs_class_posixlt: return "posixlt";
   case vctrs_class_unknown: return "unknown";
   case vctrs_class_none: return "none";
   }